requirements.txt
python-dotenv
sqlalchemy
pymysql